+++ lib/src/fine_tuned.dart
@@ -1,5 +1,8 @@
 import 'package:chat_gpt_sdk/src/model/cancel/cancel_data.dart';
 import 'package:chat_gpt_sdk/src/model/fine_tune/request/create_fine_tune.dart';
+import 'package:chat_gpt_sdk/src/model/fine_tune/request/create_fine_tune_job.dart';
+import 'package:chat_gpt_sdk/src/model/fine_tune/response/job/fine_tune_list.dart';
+import 'package:chat_gpt_sdk/src/model/fine_tune/response/job/fine_tune_model_job.dart';
 import 'package:chat_gpt_sdk/src/utils/constants.dart';
 import 'client/client.dart';
 import 'model/fine_tune/response/fine_tune_delete.dart';
@@ -21,7 +24,7 @@ class FineTuned {
       kURL + kFineTune,
       request.toJson(),
       onCancel: (it) => onCancel != null ? onCancel(it) : null,
-      onSuccess: (it) => FineTuneModel.formJson(it),
+      onSuccess: (it) => FineTuneModel.fromJson(it),
     );
   }
 
@@ -35,7 +38,7 @@ class FineTuned {
       onSuccess: (it) {
         final data = it['data'] as List;
 
-        return data.map((e) => FineTuneModel.formJson(e)).toList();
+        return data.map((e) => FineTuneModel.fromJson(e)).toList();
       },
     );
   }
@@ -48,7 +51,7 @@ class FineTuned {
     return _client.get(
       "$kURL$kFineTune/$fineTuneId",
       onCancel: (it) => onCancel != null ? onCancel(it) : null,
-      onSuccess: (it) => FineTuneModel.formJson(it),
+      onSuccess: (it) => FineTuneModel.fromJson(it),
     );
   }
 
@@ -61,7 +64,7 @@ class FineTuned {
       "$kURL$kFineTune/$fineTuneId/cancel",
       {},
       onCancel: (it) => onCancel != null ? onCancel(it) : null,
-      onSuccess: (it) => FineTuneModel.formJson(it),
+      onSuccess: (it) => FineTuneModel.fromJson(it),
     );
   }
 
@@ -88,7 +91,94 @@ class FineTuned {
       onSuccess: (it) {
         final data = it['data'] as List;
 
-        return data.map((e) => FineTuneModel.formJson(e)).toList();
+        return data.map((e) => FineTuneModel.fromJson(e)).toList();
+      },
+    );
+  }
+
+/**
+ * new api
+ */
+
+  ///Creates a job that fine-tunes a specified model from a given dataset.
+  ///Response includes details of the enqueued job including job status
+  ///and the name of the fine-tuned models once complete. [createFineTuneJob]
+  Future<FineTuneModelJob> createFineTuneJob(
+    CreateFineTuneJob request, {
+    void Function(CancelData cancelData)? onCancel,
+  }) {
+    return _client.post(
+      kURL + kFineTuneJob,
+      request.toJson(),
+      onSuccess: FineTuneModelJob.fromJson,
+      onCancel: (it) => onCancel != null ? onCancel(it) : null,
+    );
+  }
+
+  ///Get info about a fine-tuning job.[retrieveFineTuneJob]
+  Future<FineTuneList> retrieveFineTuneJob(
+    String fineTuneId, {
+    void Function(CancelData cancelData)? onCancel,
+  }) {
+    return _client.get(
+      kURL + kFineTuneJob + "/$fineTuneId",
+      onSuccess: FineTuneList.fromJson,
+      onCancel: (it) => onCancel != null ? onCancel(it) : null,
+    );
+  }
+
+  ///List your organization's fine-tuning jobs
+  Future<List<FineTuneList>> listFineTuneJob({
+    void Function(CancelData cancelData)? onCancel,
+  }) async {
+    return _client.get(
+      kURL + kFineTune,
+      onCancel: (it) => onCancel != null ? onCancel(it) : null,
+      onSuccess: (it) {
+        final data = it['data'] as List;
+
+        return data.map((e) => FineTuneList.fromJson(e)).toList();
+      },
+    );
+  }
+
+  ///Immediately cancel a fine-tune job.
+  Future<FineTuneList> cancelFineTuneJob(
+    String fineTuneId, {
+    void Function(CancelData cancelData)? onCancel,
+  }) {
+    return _client.post(
+      "$kURL$kFineTuneJob/$fineTuneId/cancel",
+      {},
+      onCancel: (it) => onCancel != null ? onCancel(it) : null,
+      onSuccess: FineTuneList.fromJson,
+    );
+  }
+
+  ///Get status updates for a fine-tuning job.
+  Stream<List<FineTuneList>> listFineTuneJobStream(
+    String fineTuneId, {
+    String? after,
+    int? limit,
+    void Function(CancelData cancelData)? onCancel,
+  }) {
+    Map<String, dynamic>? query;
+
+    if (after != null || limit != null) {
+      query = Map.of({
+        'after': after,
+        'limit': limit,
+      });
+    }
+
+    return _client.getStream(
+      "$kURL$kFineTuneJob/$fineTuneId/events",
+      queryParameters: query,
+      onCancel: (it) => onCancel != null ? onCancel(it) : null,
+      onSuccess: (it) {
+        final data = it['data'] as List;
+
+        return data.map((e) => FineTuneList.fromJson(e)).toList();
       },
     );
   }