+++ lib/src/model/fine_tune/request/create_fine_tune.dart
@@ -29,7 +29,6 @@ class CreateFineTune {
   /// To learn more about these models,
   /// see the Models documentation. [model]
   /// ## Fine models
-  /// - GptTurbo0631Model();
   /// - AdaFineModel();
   /// - BabbageFineModel();
   /// - CurieFineModel();