+++ test/openai_test.mocks.dart
@@ -3,59 +3,65 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i22;
+import 'dart:async' as _i24;
 
 import 'package:chat_gpt_sdk/src/audio.dart' as _i4;
 import 'package:chat_gpt_sdk/src/edit.dart' as _i2;
 import 'package:chat_gpt_sdk/src/embedding.dart' as _i3;
 import 'package:chat_gpt_sdk/src/fine_tuned.dart' as _i6;
 import 'package:chat_gpt_sdk/src/model/audio/request/audio_request.dart'
-    as _i30;
+    as _i32;
 import 'package:chat_gpt_sdk/src/model/audio/response/audio_response.dart'
     as _i11;
-import 'package:chat_gpt_sdk/src/model/cancel/cancel_data.dart' as _i23;
+import 'package:chat_gpt_sdk/src/model/cancel/cancel_data.dart' as _i25;
 import 'package:chat_gpt_sdk/src/model/chat_complete/request/chat_complete_text.dart'
-    as _i27;
-import 'package:chat_gpt_sdk/src/model/chat_complete/response/chat_ct_response.dart'
-    as _i26;
-import 'package:chat_gpt_sdk/src/model/chat_complete/response/chat_response_sse.dart'
     as _i29;
-import 'package:chat_gpt_sdk/src/model/client/http_setup.dart' as _i21;
+import 'package:chat_gpt_sdk/src/model/chat_complete/response/chat_ct_response.dart'
+    as _i28;
+import 'package:chat_gpt_sdk/src/model/chat_complete/response/chat_response_sse.dart'
+    as _i31;
+import 'package:chat_gpt_sdk/src/model/client/http_setup.dart' as _i23;
 import 'package:chat_gpt_sdk/src/model/complete_text/request/complete_text.dart'
-    as _i25;
+    as _i27;
 import 'package:chat_gpt_sdk/src/model/complete_text/response/complete_response.dart'
-    as _i24;
-import 'package:chat_gpt_sdk/src/model/edits/request/edit_request.dart' as _i31;
+    as _i26;
+import 'package:chat_gpt_sdk/src/model/edits/request/edit_request.dart' as _i33;
 import 'package:chat_gpt_sdk/src/model/edits/response/edit_response.dart'
     as _i12;
 import 'package:chat_gpt_sdk/src/model/embedding/request/embed_request.dart'
-    as _i34;
+    as _i36;
 import 'package:chat_gpt_sdk/src/model/embedding/response/embed_response.dart'
     as _i14;
-import 'package:chat_gpt_sdk/src/model/file/request/upload_file.dart' as _i37;
-import 'package:chat_gpt_sdk/src/model/file/response/delete_file.dart' as _i20;
+import 'package:chat_gpt_sdk/src/model/file/request/upload_file.dart' as _i40;
+import 'package:chat_gpt_sdk/src/model/file/response/delete_file.dart' as _i22;
 import 'package:chat_gpt_sdk/src/model/file/response/file_response.dart'
-    as _i18;
+    as _i20;
 import 'package:chat_gpt_sdk/src/model/file/response/upload_response.dart'
-    as _i19;
+    as _i21;
 import 'package:chat_gpt_sdk/src/model/fine_tune/request/create_fine_tune.dart'
-    as _i35;
+    as _i37;
+import 'package:chat_gpt_sdk/src/model/fine_tune/request/create_fine_tune_job.dart'
+    as _i38;
 import 'package:chat_gpt_sdk/src/model/fine_tune/response/fine_tune_delete.dart'
     as _i16;
 import 'package:chat_gpt_sdk/src/model/fine_tune/response/fine_tune_model.dart'
     as _i15;
+import 'package:chat_gpt_sdk/src/model/fine_tune/response/job/fine_tune_list.dart'
+    as _i18;
+import 'package:chat_gpt_sdk/src/model/fine_tune/response/job/fine_tune_model_job.dart'
+    as _i17;
 import 'package:chat_gpt_sdk/src/model/gen_image/request/edit_file.dart'
-    as _i32;
+    as _i34;
 import 'package:chat_gpt_sdk/src/model/gen_image/request/generate_image.dart'
-    as _i28;
+    as _i30;
 import 'package:chat_gpt_sdk/src/model/gen_image/request/variation.dart'
-    as _i33;
+    as _i35;
 import 'package:chat_gpt_sdk/src/model/gen_image/response/gen_img_response.dart'
     as _i13;
 import 'package:chat_gpt_sdk/src/model/moderation/enum/moderation_model.dart'
-    as _i36;
+    as _i39;
 import 'package:chat_gpt_sdk/src/model/moderation/response/moderation_data.dart'
-    as _i17;
+    as _i19;
 import 'package:chat_gpt_sdk/src/model/openai_engine/engine_model.dart' as _i10;
 import 'package:chat_gpt_sdk/src/model/openai_model/openai_model.dart' as _i9;
 import 'package:chat_gpt_sdk/src/moderation.dart' as _i7;
@@ -228,9 +234,9 @@ class _FakeFineTuneDelete_14 extends _i1.SmartFake
         );
 }
 
-class _FakeModerationData_15 extends _i1.SmartFake
-    implements _i17.ModerationData {
-  _FakeModerationData_15(
+class _FakeFineTuneModelJob_15 extends _i1.SmartFake
+    implements _i17.FineTuneModelJob {
+  _FakeFineTuneModelJob_15(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -239,8 +245,8 @@ class _FakeModerationData_15 extends _i1.SmartFake
         );
 }
 
-class _FakeFileResponse_16 extends _i1.SmartFake implements _i18.FileResponse {
-  _FakeFileResponse_16(
+class _FakeFineTuneList_16 extends _i1.SmartFake implements _i18.FineTuneList {
+  _FakeFineTuneList_16(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -249,9 +255,9 @@ class _FakeFileResponse_16 extends _i1.SmartFake implements _i18.FileResponse {
         );
 }
 
-class _FakeUploadResponse_17 extends _i1.SmartFake
-    implements _i19.UploadResponse {
-  _FakeUploadResponse_17(
+class _FakeModerationData_17 extends _i1.SmartFake
+    implements _i19.ModerationData {
+  _FakeModerationData_17(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -260,8 +266,29 @@ class _FakeUploadResponse_17 extends _i1.SmartFake
         );
 }
 
-class _FakeDeleteFile_18 extends _i1.SmartFake implements _i20.DeleteFile {
-  _FakeDeleteFile_18(
+class _FakeFileResponse_18 extends _i1.SmartFake implements _i20.FileResponse {
+  _FakeFileResponse_18(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeUploadResponse_19 extends _i1.SmartFake
+    implements _i21.UploadResponse {
+  _FakeUploadResponse_19(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeDeleteFile_20 extends _i1.SmartFake implements _i22.DeleteFile {
+  _FakeDeleteFile_20(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -378,7 +405,7 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
   _i8.OpenAI build({
     String? token,
     String? orgId,
-    _i21.HttpSetup? baseOption,
+    _i23.HttpSetup? baseOption,
     bool? enableLog = false,
   }) =>
       (super.noSuchMethod(
@@ -420,15 +447,15 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
         ),
       ) as _i8.OpenAI);
   @override
-  _i22.Future<_i9.OpenAiModel> listModel(
-          {void Function(_i23.CancelData)? onCancel}) =>
+  _i24.Future<_i9.OpenAiModel> listModel(
+          {void Function(_i25.CancelData)? onCancel}) =>
       (super.noSuchMethod(
         Invocation.method(
           #listModel,
           [],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i9.OpenAiModel>.value(_FakeOpenAiModel_7(
+        returnValue: _i24.Future<_i9.OpenAiModel>.value(_FakeOpenAiModel_7(
           this,
           Invocation.method(
             #listModel,
@@ -437,7 +464,7 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i9.OpenAiModel>.value(_FakeOpenAiModel_7(
+            _i24.Future<_i9.OpenAiModel>.value(_FakeOpenAiModel_7(
           this,
           Invocation.method(
             #listModel,
@@ -445,17 +472,17 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i9.OpenAiModel>);
+      ) as _i24.Future<_i9.OpenAiModel>);
   @override
-  _i22.Future<_i10.EngineModel> listEngine(
-          {void Function(_i23.CancelData)? onCancel}) =>
+  _i24.Future<_i10.EngineModel> listEngine(
+          {void Function(_i25.CancelData)? onCancel}) =>
       (super.noSuchMethod(
         Invocation.method(
           #listEngine,
           [],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i10.EngineModel>.value(_FakeEngineModel_8(
+        returnValue: _i24.Future<_i10.EngineModel>.value(_FakeEngineModel_8(
           this,
           Invocation.method(
             #listEngine,
@@ -464,7 +491,7 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i10.EngineModel>.value(_FakeEngineModel_8(
+            _i24.Future<_i10.EngineModel>.value(_FakeEngineModel_8(
           this,
           Invocation.method(
             #listEngine,
@@ -472,11 +499,11 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i10.EngineModel>);
+      ) as _i24.Future<_i10.EngineModel>);
   @override
-  _i22.Future<_i24.CompleteResponse?> onCompletion({
-    required _i25.CompleteText? request,
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i26.CompleteResponse?> onCompletion({
+    required _i27.CompleteText? request,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -487,13 +514,13 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             #onCancel: onCancel,
           },
         ),
-        returnValue: _i22.Future<_i24.CompleteResponse?>.value(),
-        returnValueForMissingStub: _i22.Future<_i24.CompleteResponse?>.value(),
-      ) as _i22.Future<_i24.CompleteResponse?>);
+        returnValue: _i24.Future<_i26.CompleteResponse?>.value(),
+        returnValueForMissingStub: _i24.Future<_i26.CompleteResponse?>.value(),
+      ) as _i24.Future<_i26.CompleteResponse?>);
   @override
-  _i22.Future<_i26.ChatCTResponse?> onChatCompletion({
-    required _i27.ChatCompleteText? request,
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i28.ChatCTResponse?> onChatCompletion({
+    required _i29.ChatCompleteText? request,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -504,13 +531,13 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             #onCancel: onCancel,
           },
         ),
-        returnValue: _i22.Future<_i26.ChatCTResponse?>.value(),
-        returnValueForMissingStub: _i22.Future<_i26.ChatCTResponse?>.value(),
-      ) as _i22.Future<_i26.ChatCTResponse?>);
+        returnValue: _i24.Future<_i28.ChatCTResponse?>.value(),
+        returnValueForMissingStub: _i24.Future<_i28.ChatCTResponse?>.value(),
+      ) as _i24.Future<_i28.ChatCTResponse?>);
   @override
-  _i22.Future<_i13.GenImgResponse?> generateImage(
-    _i28.GenerateImage? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i13.GenImgResponse?> generateImage(
+    _i30.GenerateImage? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -518,13 +545,13 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
           [request],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i13.GenImgResponse?>.value(),
-        returnValueForMissingStub: _i22.Future<_i13.GenImgResponse?>.value(),
-      ) as _i22.Future<_i13.GenImgResponse?>);
+        returnValue: _i24.Future<_i13.GenImgResponse?>.value(),
+        returnValueForMissingStub: _i24.Future<_i13.GenImgResponse?>.value(),
+      ) as _i24.Future<_i13.GenImgResponse?>);
   @override
-  _i22.Stream<_i29.ChatResponseSSE> onChatCompletionSSE({
-    required _i27.ChatCompleteText? request,
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Stream<_i31.ChatResponseSSE> onChatCompletionSSE({
+    required _i29.ChatCompleteText? request,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -535,13 +562,13 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             #onCancel: onCancel,
           },
         ),
-        returnValue: _i22.Stream<_i29.ChatResponseSSE>.empty(),
-        returnValueForMissingStub: _i22.Stream<_i29.ChatResponseSSE>.empty(),
-      ) as _i22.Stream<_i29.ChatResponseSSE>);
+        returnValue: _i24.Stream<_i31.ChatResponseSSE>.empty(),
+        returnValueForMissingStub: _i24.Stream<_i31.ChatResponseSSE>.empty(),
+      ) as _i24.Stream<_i31.ChatResponseSSE>);
   @override
-  _i22.Stream<_i24.CompleteResponse> onCompletionSSE({
-    required _i25.CompleteText? request,
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Stream<_i26.CompleteResponse> onCompletionSSE({
+    required _i27.CompleteText? request,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -552,9 +579,9 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
             #onCancel: onCancel,
           },
         ),
-        returnValue: _i22.Stream<_i24.CompleteResponse>.empty(),
-        returnValueForMissingStub: _i22.Stream<_i24.CompleteResponse>.empty(),
-      ) as _i22.Stream<_i24.CompleteResponse>);
+        returnValue: _i24.Stream<_i26.CompleteResponse>.empty(),
+        returnValueForMissingStub: _i24.Stream<_i26.CompleteResponse>.empty(),
+      ) as _i24.Stream<_i26.CompleteResponse>);
 }
 
 /// A class which mocks [Audio].
@@ -562,9 +589,9 @@ class MockOpenAI extends _i1.Mock implements _i8.OpenAI {
 /// See the documentation for Mockito's code generation for more information.
 class MockAudio extends _i1.Mock implements _i4.Audio {
   @override
-  _i22.Future<_i11.AudioResponse> transcribes(
-    _i30.AudioRequest? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i11.AudioResponse> transcribes(
+    _i32.AudioRequest? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -572,7 +599,7 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
           [request],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
+        returnValue: _i24.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
           this,
           Invocation.method(
             #transcribes,
@@ -581,7 +608,7 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
+            _i24.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
           this,
           Invocation.method(
             #transcribes,
@@ -589,11 +616,11 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i11.AudioResponse>);
+      ) as _i24.Future<_i11.AudioResponse>);
   @override
-  _i22.Future<_i11.AudioResponse> translate(
-    _i30.AudioRequest? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i11.AudioResponse> translate(
+    _i32.AudioRequest? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -601,7 +628,7 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
           [request],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
+        returnValue: _i24.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
           this,
           Invocation.method(
             #translate,
@@ -610,7 +637,7 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
+            _i24.Future<_i11.AudioResponse>.value(_FakeAudioResponse_9(
           this,
           Invocation.method(
             #translate,
@@ -618,7 +645,7 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i11.AudioResponse>);
+      ) as _i24.Future<_i11.AudioResponse>);
 }
 
 /// A class which mocks [Edit].
@@ -626,9 +653,9 @@ class MockAudio extends _i1.Mock implements _i4.Audio {
 /// See the documentation for Mockito's code generation for more information.
 class MockEdit extends _i1.Mock implements _i2.Edit {
   @override
-  _i22.Future<_i12.EditResponse> prompt(
-    _i31.EditRequest? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i12.EditResponse> prompt(
+    _i33.EditRequest? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -636,7 +663,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           [request],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i12.EditResponse>.value(_FakeEditResponse_10(
+        returnValue: _i24.Future<_i12.EditResponse>.value(_FakeEditResponse_10(
           this,
           Invocation.method(
             #prompt,
@@ -645,7 +672,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i12.EditResponse>.value(_FakeEditResponse_10(
+            _i24.Future<_i12.EditResponse>.value(_FakeEditResponse_10(
           this,
           Invocation.method(
             #prompt,
@@ -653,11 +680,11 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i12.EditResponse>);
+      ) as _i24.Future<_i12.EditResponse>);
   @override
-  _i22.Future<_i13.GenImgResponse> editImage(
-    _i32.EditImageRequest? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i13.GenImgResponse> editImage(
+    _i34.EditImageRequest? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -666,7 +693,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
+            _i24.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
           this,
           Invocation.method(
             #editImage,
@@ -675,7 +702,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
+            _i24.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
           this,
           Invocation.method(
             #editImage,
@@ -683,11 +710,11 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i13.GenImgResponse>);
+      ) as _i24.Future<_i13.GenImgResponse>);
   @override
-  _i22.Future<_i13.GenImgResponse> variation(
-    _i33.Variation? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i13.GenImgResponse> variation(
+    _i35.Variation? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -696,7 +723,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
+            _i24.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
           this,
           Invocation.method(
             #variation,
@@ -705,7 +732,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
+            _i24.Future<_i13.GenImgResponse>.value(_FakeGenImgResponse_11(
           this,
           Invocation.method(
             #variation,
@@ -713,7 +740,7 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i13.GenImgResponse>);
+      ) as _i24.Future<_i13.GenImgResponse>);
 }
 
 /// A class which mocks [Embedding].
@@ -721,9 +748,9 @@ class MockEdit extends _i1.Mock implements _i2.Edit {
 /// See the documentation for Mockito's code generation for more information.
 class MockEmbedding extends _i1.Mock implements _i3.Embedding {
   @override
-  _i22.Future<_i14.EmbedResponse> embedding(
-    _i34.EmbedRequest? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i14.EmbedResponse> embedding(
+    _i36.EmbedRequest? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -732,7 +759,7 @@ class MockEmbedding extends _i1.Mock implements _i3.Embedding {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i14.EmbedResponse>.value(_FakeEmbedResponse_12(
+            _i24.Future<_i14.EmbedResponse>.value(_FakeEmbedResponse_12(
           this,
           Invocation.method(
             #embedding,
@@ -741,7 +768,7 @@ class MockEmbedding extends _i1.Mock implements _i3.Embedding {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i14.EmbedResponse>.value(_FakeEmbedResponse_12(
+            _i24.Future<_i14.EmbedResponse>.value(_FakeEmbedResponse_12(
           this,
           Invocation.method(
             #embedding,
@@ -749,7 +776,7 @@ class MockEmbedding extends _i1.Mock implements _i3.Embedding {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i14.EmbedResponse>);
+      ) as _i24.Future<_i14.EmbedResponse>);
 }
 
 /// A class which mocks [FineTuned].
@@ -757,9 +784,9 @@ class MockEmbedding extends _i1.Mock implements _i3.Embedding {
 /// See the documentation for Mockito's code generation for more information.
 class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
   @override
-  _i22.Future<_i15.FineTuneModel> create(
-    _i35.CreateFineTune? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i15.FineTuneModel> create(
+    _i37.CreateFineTune? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -768,7 +795,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #create,
@@ -777,7 +804,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #create,
@@ -785,10 +812,10 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i15.FineTuneModel>);
+      ) as _i24.Future<_i15.FineTuneModel>);
   @override
-  _i22.Future<List<_i15.FineTuneModel>> list(
-          {void Function(_i23.CancelData)? onCancel}) =>
+  _i24.Future<List<_i15.FineTuneModel>> list(
+          {void Function(_i25.CancelData)? onCancel}) =>
       (super.noSuchMethod(
         Invocation.method(
           #list,
@@ -796,14 +823,14 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<List<_i15.FineTuneModel>>.value(<_i15.FineTuneModel>[]),
+            _i24.Future<List<_i15.FineTuneModel>>.value(<_i15.FineTuneModel>[]),
         returnValueForMissingStub:
-            _i22.Future<List<_i15.FineTuneModel>>.value(<_i15.FineTuneModel>[]),
-      ) as _i22.Future<List<_i15.FineTuneModel>>);
+            _i24.Future<List<_i15.FineTuneModel>>.value(<_i15.FineTuneModel>[]),
+      ) as _i24.Future<List<_i15.FineTuneModel>>);
   @override
-  _i22.Future<_i15.FineTuneModel> retrieve(
+  _i24.Future<_i15.FineTuneModel> retrieve(
     String? fineTuneId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -812,7 +839,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #retrieve,
@@ -821,7 +848,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #retrieve,
@@ -829,11 +856,11 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i15.FineTuneModel>);
+      ) as _i24.Future<_i15.FineTuneModel>);
   @override
-  _i22.Future<_i15.FineTuneModel> cancel(
+  _i24.Future<_i15.FineTuneModel> cancel(
     String? fineTuneId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -842,7 +869,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #cancel,
@@ -851,7 +878,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
+            _i24.Future<_i15.FineTuneModel>.value(_FakeFineTuneModel_13(
           this,
           Invocation.method(
             #cancel,
@@ -859,11 +886,11 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i15.FineTuneModel>);
+      ) as _i24.Future<_i15.FineTuneModel>);
   @override
-  _i22.Future<_i16.FineTuneDelete> delete(
+  _i24.Future<_i16.FineTuneDelete> delete(
     String? model, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -872,7 +899,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i16.FineTuneDelete>.value(_FakeFineTuneDelete_14(
+            _i24.Future<_i16.FineTuneDelete>.value(_FakeFineTuneDelete_14(
           this,
           Invocation.method(
             #delete,
@@ -881,7 +908,7 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i16.FineTuneDelete>.value(_FakeFineTuneDelete_14(
+            _i24.Future<_i16.FineTuneDelete>.value(_FakeFineTuneDelete_14(
           this,
           Invocation.method(
             #delete,
@@ -889,11 +916,11 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i16.FineTuneDelete>);
+      ) as _i24.Future<_i16.FineTuneDelete>);
   @override
-  _i22.Stream<List<_i15.FineTuneModel>> listStream(
+  _i24.Stream<List<_i15.FineTuneModel>> listStream(
     String? fineTuneId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -901,10 +928,132 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
           [fineTuneId],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Stream<List<_i15.FineTuneModel>>.empty(),
+        returnValue: _i24.Stream<List<_i15.FineTuneModel>>.empty(),
         returnValueForMissingStub:
-            _i22.Stream<List<_i15.FineTuneModel>>.empty(),
-      ) as _i22.Stream<List<_i15.FineTuneModel>>);
+            _i24.Stream<List<_i15.FineTuneModel>>.empty(),
+      ) as _i24.Stream<List<_i15.FineTuneModel>>);
+  @override
+  _i24.Future<_i17.FineTuneModelJob> createFineTuneJob(
+    _i38.CreateFineTuneJob? request, {
+    void Function(_i25.CancelData)? onCancel,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #createFineTuneJob,
+          [request],
+          {#onCancel: onCancel},
+        ),
+        returnValue:
+            _i24.Future<_i17.FineTuneModelJob>.value(_FakeFineTuneModelJob_15(
+          this,
+          Invocation.method(
+            #createFineTuneJob,
+            [request],
+            {#onCancel: onCancel},
+          ),
+        )),
+        returnValueForMissingStub:
+            _i24.Future<_i17.FineTuneModelJob>.value(_FakeFineTuneModelJob_15(
+          this,
+          Invocation.method(
+            #createFineTuneJob,
+            [request],
+            {#onCancel: onCancel},
+          ),
+        )),
+      ) as _i24.Future<_i17.FineTuneModelJob>);
+  @override
+  _i24.Future<_i18.FineTuneList> retrieveFineTuneJob(
+    String? fineTuneId, {
+    void Function(_i25.CancelData)? onCancel,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #retrieveFineTuneJob,
+          [fineTuneId],
+          {#onCancel: onCancel},
+        ),
+        returnValue: _i24.Future<_i18.FineTuneList>.value(_FakeFineTuneList_16(
+          this,
+          Invocation.method(
+            #retrieveFineTuneJob,
+            [fineTuneId],
+            {#onCancel: onCancel},
+          ),
+        )),
+        returnValueForMissingStub:
+            _i24.Future<_i18.FineTuneList>.value(_FakeFineTuneList_16(
+          this,
+          Invocation.method(
+            #retrieveFineTuneJob,
+            [fineTuneId],
+            {#onCancel: onCancel},
+          ),
+        )),
+      ) as _i24.Future<_i18.FineTuneList>);
+  @override
+  _i24.Future<List<_i18.FineTuneList>> listFineTuneJob(
+          {void Function(_i25.CancelData)? onCancel}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #listFineTuneJob,
+          [],
+          {#onCancel: onCancel},
+        ),
+        returnValue:
+            _i24.Future<List<_i18.FineTuneList>>.value(<_i18.FineTuneList>[]),
+        returnValueForMissingStub:
+            _i24.Future<List<_i18.FineTuneList>>.value(<_i18.FineTuneList>[]),
+      ) as _i24.Future<List<_i18.FineTuneList>>);
+  @override
+  _i24.Future<_i18.FineTuneList> cancelFineTuneJob(
+    String? fineTuneId, {
+    void Function(_i25.CancelData)? onCancel,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #cancelFineTuneJob,
+          [fineTuneId],
+          {#onCancel: onCancel},
+        ),
+        returnValue: _i24.Future<_i18.FineTuneList>.value(_FakeFineTuneList_16(
+          this,
+          Invocation.method(
+            #cancelFineTuneJob,
+            [fineTuneId],
+            {#onCancel: onCancel},
+          ),
+        )),
+        returnValueForMissingStub:
+            _i24.Future<_i18.FineTuneList>.value(_FakeFineTuneList_16(
+          this,
+          Invocation.method(
+            #cancelFineTuneJob,
+            [fineTuneId],
+            {#onCancel: onCancel},
+          ),
+        )),
+      ) as _i24.Future<_i18.FineTuneList>);
+  @override
+  _i24.Stream<List<_i18.FineTuneList>> listFineTuneJobStream(
+    String? fineTuneId, {
+    String? after,
+    int? limit,
+    void Function(_i25.CancelData)? onCancel,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #listFineTuneJobStream,
+          [fineTuneId],
+          {
+            #after: after,
+            #limit: limit,
+            #onCancel: onCancel,
+          },
+        ),
+        returnValue: _i24.Stream<List<_i18.FineTuneList>>.empty(),
+        returnValueForMissingStub: _i24.Stream<List<_i18.FineTuneList>>.empty(),
+      ) as _i24.Stream<List<_i18.FineTuneList>>);
 }
 
 /// A class which mocks [Moderation].
@@ -912,10 +1061,10 @@ class MockFineTuned extends _i1.Mock implements _i6.FineTuned {
 /// See the documentation for Mockito's code generation for more information.
 class MockModeration extends _i1.Mock implements _i7.Moderation {
   @override
-  _i22.Future<_i17.ModerationData> create({
+  _i24.Future<_i19.ModerationData> create({
     required String? input,
-    required _i36.ModerationModel? model,
-    void Function(_i23.CancelData)? onCancel,
+    required _i39.ModerationModel? model,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -928,7 +1077,7 @@ class MockModeration extends _i1.Mock implements _i7.Moderation {
           },
         ),
         returnValue:
-            _i22.Future<_i17.ModerationData>.value(_FakeModerationData_15(
+            _i24.Future<_i19.ModerationData>.value(_FakeModerationData_17(
           this,
           Invocation.method(
             #create,
@@ -941,7 +1090,7 @@ class MockModeration extends _i1.Mock implements _i7.Moderation {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i17.ModerationData>.value(_FakeModerationData_15(
+            _i24.Future<_i19.ModerationData>.value(_FakeModerationData_17(
           this,
           Invocation.method(
             #create,
@@ -953,7 +1102,7 @@ class MockModeration extends _i1.Mock implements _i7.Moderation {
             },
           ),
         )),
-      ) as _i22.Future<_i17.ModerationData>);
+      ) as _i24.Future<_i19.ModerationData>);
 }
 
 /// A class which mocks [OpenAIFile].
@@ -961,15 +1110,15 @@ class MockModeration extends _i1.Mock implements _i7.Moderation {
 /// See the documentation for Mockito's code generation for more information.
 class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
   @override
-  _i22.Future<_i18.FileResponse> get(
-          {void Function(_i23.CancelData)? onCancel}) =>
+  _i24.Future<_i20.FileResponse> get(
+          {void Function(_i25.CancelData)? onCancel}) =>
       (super.noSuchMethod(
         Invocation.method(
           #get,
           [],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i18.FileResponse>.value(_FakeFileResponse_16(
+        returnValue: _i24.Future<_i20.FileResponse>.value(_FakeFileResponse_18(
           this,
           Invocation.method(
             #get,
@@ -978,7 +1127,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i18.FileResponse>.value(_FakeFileResponse_16(
+            _i24.Future<_i20.FileResponse>.value(_FakeFileResponse_18(
           this,
           Invocation.method(
             #get,
@@ -986,11 +1135,11 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i18.FileResponse>);
+      ) as _i24.Future<_i20.FileResponse>);
   @override
-  _i22.Future<_i19.UploadResponse> uploadFile(
-    _i37.UploadFile? request, {
-    void Function(_i23.CancelData)? onCancel,
+  _i24.Future<_i21.UploadResponse> uploadFile(
+    _i40.UploadFile? request, {
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -999,7 +1148,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i19.UploadResponse>.value(_FakeUploadResponse_17(
+            _i24.Future<_i21.UploadResponse>.value(_FakeUploadResponse_19(
           this,
           Invocation.method(
             #uploadFile,
@@ -1008,7 +1157,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i19.UploadResponse>.value(_FakeUploadResponse_17(
+            _i24.Future<_i21.UploadResponse>.value(_FakeUploadResponse_19(
           this,
           Invocation.method(
             #uploadFile,
@@ -1016,11 +1165,11 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i19.UploadResponse>);
+      ) as _i24.Future<_i21.UploadResponse>);
   @override
-  _i22.Future<_i20.DeleteFile> delete(
+  _i24.Future<_i22.DeleteFile> delete(
     String? fileId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1028,7 +1177,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           [fileId],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<_i20.DeleteFile>.value(_FakeDeleteFile_18(
+        returnValue: _i24.Future<_i22.DeleteFile>.value(_FakeDeleteFile_20(
           this,
           Invocation.method(
             #delete,
@@ -1037,7 +1186,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i20.DeleteFile>.value(_FakeDeleteFile_18(
+            _i24.Future<_i22.DeleteFile>.value(_FakeDeleteFile_20(
           this,
           Invocation.method(
             #delete,
@@ -1045,11 +1194,11 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i20.DeleteFile>);
+      ) as _i24.Future<_i22.DeleteFile>);
   @override
-  _i22.Future<_i19.UploadResponse> retrieve(
+  _i24.Future<_i21.UploadResponse> retrieve(
     String? fileId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1058,7 +1207,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           {#onCancel: onCancel},
         ),
         returnValue:
-            _i22.Future<_i19.UploadResponse>.value(_FakeUploadResponse_17(
+            _i24.Future<_i21.UploadResponse>.value(_FakeUploadResponse_19(
           this,
           Invocation.method(
             #retrieve,
@@ -1067,7 +1216,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           ),
         )),
         returnValueForMissingStub:
-            _i22.Future<_i19.UploadResponse>.value(_FakeUploadResponse_17(
+            _i24.Future<_i21.UploadResponse>.value(_FakeUploadResponse_19(
           this,
           Invocation.method(
             #retrieve,
@@ -1075,11 +1224,11 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
             {#onCancel: onCancel},
           ),
         )),
-      ) as _i22.Future<_i19.UploadResponse>);
+      ) as _i24.Future<_i21.UploadResponse>);
   @override
-  _i22.Future<dynamic> retrieveContent(
+  _i24.Future<dynamic> retrieveContent(
     String? fileId, {
-    void Function(_i23.CancelData)? onCancel,
+    void Function(_i25.CancelData)? onCancel,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -1087,7 +1236,7 @@ class MockOpenAIFile extends _i1.Mock implements _i5.OpenAIFile {
           [fileId],
           {#onCancel: onCancel},
         ),
-        returnValue: _i22.Future<dynamic>.value(),
-        returnValueForMissingStub: _i22.Future<dynamic>.value(),
-      ) as _i22.Future<dynamic>);
+        returnValue: _i24.Future<dynamic>.value(),
+        returnValueForMissingStub: _i24.Future<dynamic>.value(),
+      ) as _i24.Future<dynamic>);
 }